Refactored Problem class into GroupedProblem and UngroupedProblem (#681)
* Refactored Problem class into GroupedProblem and UngroupedProblem
* Added docstrings
